new prototype scene
I added the new models to the prototype, scaled it to fit with the current player/camera. I removed the old point A and point B, as well as the ship, but kept the debris and the ship door.

There is a new door. I need to figure out how to animate it (either exporting the animation, or triggering animation - I do love animating in Unity so maybe I'll do that with both the box and the door... I need to learn how to set up animation condition trees anyway.

I did move the gravity on/off thingies to fit the new scene. I also assigned box coliders to the floors, container box, and the new doors. I'm still figuring out how to do the coliders for the ship and station (which are one big piece of geometry each).

I also assigned a mesh colider to the new debris bits I threw in there.

Lemme know if I fucked anything up.